More test stuff and beginning of improving a get_object somewhere
